roll back more unnecessary renaming
